+++ CHANGES.md
@@ -63,6 +63,7 @@ Changes in 12.5
 - JS remove inline code from ChartjsChart() in Charts.fnc.php & ChartjsChart.js
 - JS remove inline code from GetFieldsForm() in Fields.fnc.php & FieldsGetForm.js
 - JS remove inline code from SubstitutionsInput() in Substitutions.fnc.php & SubstitutionsInput.js
+- JS fix CSP error: use setInnerHTML() instead of jQuery .append() function in jquery.colorbox.js & jquery.colorbox-min.js
 
 Changes in 12.4.3
 -----------------